Add SplitString functionality and tests
This commit adds the ability to split a string into a slice of
'StringFragment' structs. Each StringFragment contains a string and a
boolean describing whether that string is intended to be shuffled. In
this way, words (which will be shuffled) are distinguished from
whitespace and punctuation (which will not be shuffled).